Move the Git-source tests into their own file
`knowledge/git_source.py` became its own module, but its tests stayed in
`test_knowledge_manager.py` — a 9056-line, 237-test file whose four longest
tests were all Git tests. This gives that module's contract its own file.

Moved verbatim into `tests/test_knowledge_git_source.py`: the 16 tests whose
subject is the checkout itself — `_sync_once`, `_ensure_repository`,
`_hydrate_lfs_worktree`, `_run_git` timeout/cancellation/redaction, and the
credential-injection tests proving secrets reach `git` only through
process-local `GIT_CONFIG_*` and never land in the checkout's config or in
published metadata. The `_git_manager` helper moved with them; it had no other
callers.

Tests where Git is only the source fixture stay put: the no-op refresh fast
path, incremental reindex cost, duplicate-root config validation, refresh
cooldowns and polling, and the `redaction.py` URL unit tests.

The in-memory Chroma fake, the autouse `patch_vector_store` fixture and the
`_config` factory move to `tests/knowledge_test_support.py` so both files share
one definition rather than duplicating it.

Collected tests: 237 before, 220 + 17 after.
