+++ tests/test_knowledge_manager.py
@@ -11,16 +11,13 @@ import sys
 from contextlib import asynccontextmanager
 from dataclasses import dataclass, replace
 from datetime import UTC, datetime
-from itertools import count
 from pathlib import Path
-from threading import Event, Lock, get_ident
-from typing import TYPE_CHECKING, ClassVar, cast
+from threading import Event, get_ident
+from typing import TYPE_CHECKING, cast
 from unittest.mock import MagicMock
 
 import httpx
 import pytest
-from agno.knowledge.document.base import Document
-from agno.knowledge.embedder.base import Embedder
 from fastapi.testclient import TestClient
 from openai import AuthenticationError
 from pydantic import ValidationError
@@ -76,182 +73,23 @@ from mindroom.knowledge.utils import KnowledgeAvailabilityDetail
 from mindroom.knowledge.watch import KnowledgeSourceWatcher
 from mindroom.tool_system.worker_routing import ToolExecutionIdentity
 from tests.conftest import bind_runtime_paths, runtime_paths_for, test_runtime_paths
-from tests.knowledge_test_support import chroma_get_result, metadata_matches
+from tests.knowledge_test_support import (
+    _Client,
+    _Collection,
+    _config,
+    _Knowledge,
+    _vector_row_ids,
+    _VectorDb,
+    patch_vector_store,  # noqa: F401  # autouse fixture
+)
 
 if TYPE_CHECKING:
-    from collections.abc import AsyncIterator, Coroutine, Iterable, Iterator, Sequence
+    from collections.abc import AsyncIterator, Coroutine, Iterable
     from types import ModuleType
 
     from agno.knowledge.reader.base import Reader
 
     from mindroom.constants import RuntimePaths
-
-
-_vector_row_ids = count()
-
-
-class _Collection:
-    def __init__(self, name: str) -> None:
-        self._name = name
-
-    def get(
-        self,
-        *,
-        include: Sequence[str],
-        limit: int | None = None,
-        offset: int = 0,
-        where: dict[str, object] | None = None,
-    ) -> dict[str, object]:
-        with _VectorDb.lock:
-            selected_all = list(_VectorDb.collections.get(self._name, []))
-        if where:
-            key, condition = next(iter(where.items()))
-            selected_all = [item for item in selected_all if metadata_matches(item["metadata"], key, condition)]
-        selected = selected_all[offset:] if limit is None else selected_all[offset : offset + limit]
-        return chroma_get_result(
-            ids=[str(item["id"]) for item in selected],
-            metadatas=[dict(item["metadata"]) for item in selected],
-            documents=[str(item["content"]) for item in selected],
-            embeddings=[list(cast("list[float]", item["embedding"])) for item in selected],
-            include=include,
-        )
-
-    def add(
-        self,
-        *,
-        ids: list[str],
-        embeddings: list[list[float]],
-        documents: list[str],
-        metadatas: list[dict[str, object]],
-    ) -> None:
-        if not ids:
-            # Chroma rejects an empty write rather than treating it as a no-op.
-            message = "Expected Embeddings to be non-empty list or numpy array, got [] in add."
-            raise ValueError(message)
-        with _VectorDb.lock:
-            _VectorDb.collections.setdefault(self._name, []).extend(
-                {
-                    "id": identifier,
-                    "content": document,
-                    "embedding": list(embedding),
-                    "metadata": dict(metadata),
-                }
-                for identifier, embedding, document, metadata in zip(
-                    ids,
-                    embeddings,
-                    documents,
-                    metadatas,
-                    strict=True,
-                )
-            )
-
-    def delete(self, *, where: dict[str, object]) -> None:
-        key, condition = next(iter(where.items()))
-        with _VectorDb.lock:
-            _VectorDb.collections[self._name] = [
-                item
-                for item in _VectorDb.collections.get(self._name, [])
-                if not metadata_matches(item["metadata"], key, condition)
-            ]
-
-
-class _Client:
-    def get_collection(self, name: str) -> _Collection:
-        return _Collection(name)
-
-    def list_collections(self) -> list[str]:
-        with _VectorDb.lock:
-            return sorted(_VectorDb.collections)
-
-
-class _VectorDb:
-    collections: ClassVar[dict[str, list[dict[str, object]]]] = {}
-    lock: ClassVar[Lock] = Lock()
-
-    def __init__(self, *, collection: str, **_: object) -> None:
-        self.collection_name = collection
-        self.client = _Client()
-
-    def delete(self) -> bool:
-        with self.lock:
-            self.collections.pop(self.collection_name, None)
-        return True
-
-    def create(self) -> None:
-        with self.lock:
-            self.collections[self.collection_name] = []
-
-    def exists(self) -> bool:
-        with self.lock:
-            return self.collection_name in self.collections
-
-    def search(
-        self,
-        *,
-        query: str,
-        limit: int,
-        filters: dict[str, object] | list[object] | None = None,
-    ) -> list[Document]:
-        _ = (query, filters)
-        with self.lock:
-            items = list(self.collections.get(self.collection_name, []))
-        return [Document(content=str(item["content"]), meta_data=dict(item["metadata"])) for item in items[:limit]]
-
-    async def async_search(
-        self,
-        *,
-        query: str,
-        limit: int,
-        filters: dict[str, object] | list[object] | None = None,
-    ) -> list[Document]:
-        return self.search(query=query, limit=limit, filters=filters)
-
-
-class _Knowledge:
-    def __init__(self, vector_db: _VectorDb | None = None) -> None:
-        self.vector_db = vector_db
-
-    def insert(
-        self,
-        *,
-        path: str,
-        metadata: dict[str, object],
-        upsert: bool,
-        reader: object | None = None,
-    ) -> None:
-        _ = (upsert, reader)
-        with _VectorDb.lock:
-            _VectorDb.collections.setdefault(self.vector_db.collection_name, []).append(
-                {
-                    "id": f"row-{next(_vector_row_ids)}",
-                    "content": Path(path).read_text(encoding="utf-8"),
-                    "embedding": [1.0],
-                    "metadata": dict(metadata),
-                },
-            )
-
-    async def ainsert(
-        self,
-        *,
-        path: str,
-        metadata: dict[str, object],
-        upsert: bool,
-        reader: object | None = None,
-    ) -> None:
-        # Match the real Knowledge surface: ainsert delegates to insert.
-        self.insert(path=path, metadata=metadata, upsert=upsert, reader=reader)
-
-    def remove_vectors_by_metadata(self, metadata: dict[str, object]) -> bool:
-        with _VectorDb.lock:
-            items = _VectorDb.collections.get(self.vector_db.collection_name, [])
-            filtered = [
-                item for item in items if not all(item["metadata"].get(key) == value for key, value in metadata.items())
-            ]
-            _VectorDb.collections[self.vector_db.collection_name] = filtered
-        return len(filtered) != len(items)
-
-    def search(self, query: str, max_results: int | None = None) -> list[Document]:
-        return self.vector_db.search(query=query, limit=max_results or 5)
 
 
 def _insert_with_real_reader(
@@ -278,53 +116,11 @@ def _insert_with_real_reader(
         )
 
 
-class _FakeEmbedder(Embedder):
-    """Typed stand-in for the configured embedder; never issues requests."""
-
-    def get_embedding(self, text: str) -> list[float]:
-        _ = text
-        msg = "fake embedder must not be asked for vectors"
-        raise AssertionError(msg)
-
-    def get_embedding_and_usage(self, text: str) -> tuple[list[float], dict[str, object] | None]:
-        _ = text
-        msg = "fake embedder must not be asked for vectors"
-        raise AssertionError(msg)
-
-
 class _AutoCreatingKnowledge(_Knowledge):
     def __init__(self, vector_db: _VectorDb) -> None:
         super().__init__(vector_db)
         if not vector_db.exists():
             vector_db.create()
-
-
-@pytest.fixture(autouse=True)
-def patch_vector_store(monkeypatch: pytest.MonkeyPatch) -> Iterator[None]:
-    """Use an in-memory vector store for published knowledge index tests."""
-    _VectorDb.collections = {}
-    monkeypatch.setattr("mindroom.knowledge.manager.ChromaDb", _VectorDb)
-    monkeypatch.setattr("mindroom.knowledge.collections.ChromaDb", _VectorDb)
-    monkeypatch.setattr("mindroom.knowledge.manager.Knowledge", _Knowledge)
-    monkeypatch.setattr("mindroom.knowledge.collections.Knowledge", _Knowledge)
-    monkeypatch.setattr(
-        "mindroom.knowledge.manager.create_configured_embedder",
-        lambda *_args, **_kwargs: _FakeEmbedder(),
-    )
-    monkeypatch.setattr("mindroom.knowledge.indexing_config.ChromaDb", _VectorDb)
-    monkeypatch.setattr("mindroom.knowledge.registry.ChromaDb", _VectorDb)
-    monkeypatch.setattr("mindroom.knowledge.registry.StrictSearchKnowledge", _Knowledge)
-    monkeypatch.setattr("mindroom.knowledge.registry.create_configured_embedder", lambda *_args, **_kwargs: object())
-    knowledge_registry._published_indexes.clear()
-    knowledge_utils._refresh_scheduled_at.clear()
-    knowledge_refresh_runner._refresh_locks.clear()
-    knowledge_refresh_runner._active_refresh_counts.clear()
-    yield
-    knowledge_registry._published_indexes.clear()
-    knowledge_utils._refresh_scheduled_at.clear()
-    knowledge_refresh_runner._refresh_locks.clear()
-    knowledge_refresh_runner._active_refresh_counts.clear()
-    _VectorDb.collections = {}
 
 
 async def _wait_for_refresh_lock_borrowers(
@@ -374,36 +170,6 @@ def _test_indexing_settings(base_id: str = "docs") -> IndexingSettings:
         include_extensions="",
         exclude_extensions="()",
         extra_extensions="()",
-    )
-
-
-def _config(
-    tmp_path: Path,
-    *,
-    bases: dict[str, Path],
-    agent_bases: list[str],
-    git_configs: dict[str, KnowledgeGitConfig] | None = None,
-    watch: bool = False,
-    modes: dict[str, str] | None = None,
-    memory: dict[str, object] | None = None,
-) -> Config:
-    runtime_paths = test_runtime_paths(tmp_path)
-    return bind_runtime_paths(
-        Config(
-            agents={"helper": AgentConfig(display_name="Helper", knowledge_bases=agent_bases)},
-            models={},
-            memory=memory or {},
-            knowledge_bases={
-                base_id: KnowledgeBaseConfig(
-                    path=str(path),
-                    watch=watch,
-                    git=(git_configs or {}).get(base_id),
-                    mode=(modes or {}).get(base_id, "semantic"),
-                )
-                for base_id, path in bases.items()
-            },
-        ),
-        runtime_paths,
     )
 
 
@@ -459,32 +225,6 @@ def _record_git_sync(
     source._last_synced_head = result.head
     source._tracked_relative_paths = set(relative_paths)
     return result
-
-
-def _git_manager(
-    tmp_path: Path,
-    *,
-    lfs: bool = False,
-    include_extensions: list[str] | None = None,
-    sync_timeout_seconds: int = 3600,
-) -> KnowledgeManager:
-    knowledge_path = tmp_path / "knowledge"
-    config = _config(
-        tmp_path,
-        bases={"docs": knowledge_path},
-        agent_bases=["docs"],
-        git_configs={
-            "docs": KnowledgeGitConfig(
-                repo_url="https://example.com/org/repo.git",
-                branch="main",
-                lfs=lfs,
-                sync_timeout_seconds=sync_timeout_seconds,
-            ),
-        },
-    )
-    if include_extensions is not None:
-        config.knowledge_bases["docs"].include_extensions = include_extensions
-    return KnowledgeManager("docs", config=config, runtime_paths=runtime_paths_for(config))
 
 
 def test_cold_git_status_with_existing_non_checkout_dir_returns_empty_files(tmp_path: Path) -> None:
@@ -3157,45 +2897,6 @@ async def test_private_git_updated_refresh_preserves_execution_identity(
     assert [document.content for document in lookup.index.knowledge.search("updated", max_results=5)] == [
         "alice private git updated",
     ]
-
-
-@pytest.mark.asyncio
-async def test_git_source_sync_does_not_mutate_index_directly(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Git source sync should never bypass candidate publish by mutating the live index."""
-    docs_path = tmp_path / "docs"
-    docs_path.mkdir()
-    git_config = KnowledgeGitConfig(repo_url="https://example.com/org/repo.git", branch="main")
-    config = _config(
-        tmp_path,
-        bases={"docs": docs_path},
-        agent_bases=["docs"],
-        git_configs={"docs": git_config},
-    )
-    runtime_paths = runtime_paths_for(config)
-    manager = KnowledgeManager("docs", config, runtime_paths)
-
-    async def _sync_once(_git_config: KnowledgeGitConfig) -> tuple[set[str], set[str], bool]:
-        return {"changed.md"}, {"removed.md"}, True
-
-    async def _git_rev_parse(_ref: str) -> str:
-        return "rev-source-only"
-
-    async def _git_checkout_present() -> bool:
-        return True
-
-    monkeypatch.setattr(manager.git_source, "_sync_once", _sync_once)
-    monkeypatch.setattr(manager.git_source, "_rev_parse", _git_rev_parse)
-    monkeypatch.setattr(manager.git_source, "_checkout_present", _git_checkout_present)
-
-    result = await manager.git_source.sync()
-
-    assert not hasattr(manager, "remove_file")
-    assert not hasattr(manager, "index_file")
-    assert result == GitSyncResult(head="rev-source-only", updated=True)
-    assert manager.git_source.last_synced_head == "rev-source-only"
 
 
 @pytest.mark.asyncio
@@ -7796,307 +7497,6 @@ async def test_git_refresh_marks_duplicate_source_sibling_stale(
 
 
 @pytest.mark.asyncio
-async def test_git_credentials_service_token_stays_out_of_git_config_and_metadata(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """CredentialsManager Git secrets should be process-local, not copied into checkout config."""
-    docs_path = tmp_path / "docs"
-    git_config = KnowledgeGitConfig(
-        repo_url="https://example.com/org/private.git",
-        branch="main",
-        credentials_service="github_private",
-    )
-    config = _config(
-        tmp_path,
-        bases={"docs": docs_path},
-        agent_bases=["docs"],
-        git_configs={"docs": git_config},
-    )
-    runtime_paths = runtime_paths_for(config)
-    get_runtime_shared_credentials_manager(runtime_paths).save_credentials(
-        "github_private",
-        {"token": "secret-token"},
-    )
-    clone_envs: list[dict[str, str] | None] = []
-    clean_url = "https://example.com/org/private.git"
-
-    async def _fake_run_git(
-        self: KnowledgeManager,
-        args: list[str],
-        *,
-        cwd: Path | None = None,
-        env: dict[str, str] | None = None,
-    ) -> str:
-        _ = self
-        if args[0] == "clone":
-            clone_envs.append(env)
-            assert args[-2] == clean_url
-            target = Path(args[-1])
-            target.mkdir(parents=True, exist_ok=True)
-            await asyncio.to_thread(
-                subprocess.run,
-                ["git", "init"],
-                cwd=target,
-                check=True,
-                capture_output=True,
-                text=True,
-            )
-            await asyncio.to_thread(
-                subprocess.run,
-                ["git", "remote", "add", "origin", args[-2]],
-                cwd=target,
-                check=True,
-                capture_output=True,
-                text=True,
-            )
-            (target / "doc.md").write_text("credential service content", encoding="utf-8")
-            return ""
-        if args == ["remote", "set-url", "origin", clean_url]:
-            assert cwd is not None
-            await asyncio.to_thread(
-                subprocess.run,
-                ["git", *args],
-                cwd=cwd,
-                check=True,
-                capture_output=True,
-                text=True,
-            )
-            return ""
-        if args == ["ls-files", "-z"]:
-            return "doc.md\x00"
-        if args == ["rev-parse", "HEAD"]:
-            return "rev-auth\n"
-        return ""
-
-    monkeypatch.setattr(GitKnowledgeSource, "_run_git", _fake_run_git)
-
-    result = await refresh_knowledge_binding("docs", config=config, runtime_paths=runtime_paths)
-    key = resolve_published_index_key("docs", config=config, runtime_paths=runtime_paths)
-    metadata_text = published_index_metadata_path(key).read_text(encoding="utf-8")
-    git_config_text = (docs_path / ".git" / "config").read_text(encoding="utf-8")
-    clone_env = clone_envs[0]
-
-    assert result.index_published is True
-    assert clone_env is not None
-    assert clone_env["GIT_CONFIG_KEY_0"] == f"http.{clean_url}.extraHeader"
-    assert clone_env["GIT_CONFIG_VALUE_0"].startswith("Authorization: Basic ")
-    assert "secret-token" not in str(clone_env)
-    assert "secret-token" not in git_config_text
-    assert "x-access-token" not in git_config_text
-    assert clean_url in git_config_text
-    assert "secret-token" not in metadata_text
-    assert "x-access-token" not in metadata_text
-
-
-@pytest.mark.asyncio
-async def test_git_embedded_userinfo_url_is_not_reused_in_git_auth_env(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Embedded Git URL userinfo should become process-local auth without echoing the raw URL."""
-    docs_path = tmp_path / "docs"
-    raw_url = "https://git-user:secret-token@example.com/org/private.git"
-    clean_url = "https://example.com/org/private.git"
-    git_config = KnowledgeGitConfig(repo_url=raw_url, branch="main")
-    config = _config(
-        tmp_path,
-        bases={"docs": docs_path},
-        agent_bases=["docs"],
-        git_configs={"docs": git_config},
-    )
-    runtime_paths = runtime_paths_for(config)
-    clone_envs: list[dict[str, str] | None] = []
-
-    async def _fake_run_git(
-        self: KnowledgeManager,
-        args: list[str],
-        *,
-        cwd: Path | None = None,
-        env: dict[str, str] | None = None,
-    ) -> str:
-        _ = (self, cwd)
-        if args[0] == "clone":
-            clone_envs.append(env)
-            assert args[-2] == clean_url
-            target = Path(args[-1])
-            target.mkdir(parents=True, exist_ok=True)
-            (target / "doc.md").write_text("embedded userinfo content", encoding="utf-8")
-            return ""
-        if args == ["remote", "set-url", "origin", clean_url]:
-            return ""
-        if args == ["ls-files", "-z"]:
-            return "doc.md\x00"
-        if args == ["rev-parse", "HEAD"]:
-            return "rev-userinfo\n"
-        return ""
-
-    monkeypatch.setattr(GitKnowledgeSource, "_run_git", _fake_run_git)
-
-    result = await refresh_knowledge_binding("docs", config=config, runtime_paths=runtime_paths)
-    clone_env = clone_envs[0]
-
-    assert result.index_published is True
-    assert clone_env is not None
-    assert clone_env["GIT_CONFIG_KEY_0"] == f"http.{clean_url}.extraHeader"
-    assert clone_env["GIT_CONFIG_VALUE_0"].startswith("Authorization: Basic ")
-    assert raw_url not in str(clone_env)
-    assert "secret-token" not in str(clone_env)
-
-
-@pytest.mark.parametrize(
-    ("raw_url", "clean_url"),
-    [
-        ("ssh://git-user:secret-token@example.com/org/private.git", "ssh://example.com/org/private.git"),
-        ("git+https://git-user:secret-token@example.com/org/private.git", "git+https://example.com/org/private.git"),
-    ],
-)
-@pytest.mark.asyncio
-async def test_git_unsupported_scheme_userinfo_is_not_copied_to_git_config_env(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    raw_url: str,
-    clean_url: str,
-) -> None:
-    """Unsupported embedded userinfo must not be copied into transient Git config."""
-    docs_path = tmp_path / "docs"
-    git_config = KnowledgeGitConfig(repo_url=raw_url, branch="main")
-    config = _config(
-        tmp_path,
-        bases={"docs": docs_path},
-        agent_bases=["docs"],
-        git_configs={"docs": git_config},
-    )
-    runtime_paths = runtime_paths_for(config)
-    clone_calls: list[tuple[list[str], dict[str, str] | None]] = []
-
-    async def _fake_run_git(
-        self: KnowledgeManager,
-        args: list[str],
-        *,
-        cwd: Path | None = None,
-        env: dict[str, str] | None = None,
-    ) -> str:
-        _ = (self, cwd)
-        if args[0] == "clone":
-            clone_calls.append((list(args), env))
-            assert args[-2] == clean_url
-            target = Path(args[-1])
-            target.mkdir(parents=True, exist_ok=True)
-            (target / "doc.md").write_text("unsupported scheme userinfo content", encoding="utf-8")
-            return ""
-        if args == ["remote", "set-url", "origin", clean_url]:
-            return ""
-        if args == ["ls-files", "-z"]:
-            return "doc.md\x00"
-        if args == ["rev-parse", "HEAD"]:
-            return "rev-unsupported-userinfo\n"
-        return ""
-
-    monkeypatch.setattr(GitKnowledgeSource, "_run_git", _fake_run_git)
-
-    result = await refresh_knowledge_binding("docs", config=config, runtime_paths=runtime_paths)
-    key = resolve_published_index_key("docs", config=config, runtime_paths=runtime_paths)
-    metadata_text = published_index_metadata_path(key).read_text(encoding="utf-8")
-    clone_args, clone_env = clone_calls[0]
-    serialized_clone_call = json.dumps({"args": clone_args, "env": clone_env}, sort_keys=True)
-
-    assert result.index_published is True
-    assert clone_env is None
-    assert clean_url in clone_args
-    assert raw_url not in serialized_clone_call
-    assert "secret-token" not in serialized_clone_call
-    assert raw_url not in metadata_text
-    assert "secret-token" not in metadata_text
-
-
-@pytest.mark.asyncio
-async def test_git_query_and_fragment_tokens_stay_out_of_persistent_remote_and_metadata(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """URL query and fragment secrets should be transient auth only, never persisted."""
-    docs_path = tmp_path / "docs"
-    raw_url = "https://example.com/org/private.git?token=query-secret#frag-secret"
-    clean_url = "https://example.com/org/private.git"
-    git_config = KnowledgeGitConfig(repo_url=raw_url, branch="main")
-    config = _config(
-        tmp_path,
-        bases={"docs": docs_path},
-        agent_bases=["docs"],
-        git_configs={"docs": git_config},
-    )
-    runtime_paths = runtime_paths_for(config)
-    clone_envs: list[dict[str, str] | None] = []
-
-    async def _fake_run_git(
-        self: KnowledgeManager,
-        args: list[str],
-        *,
-        cwd: Path | None = None,
-        env: dict[str, str] | None = None,
-    ) -> str:
-        _ = self
-        if args[0] == "clone":
-            clone_envs.append(env)
-            assert args[-2] == clean_url
-            target = Path(args[-1])
-            target.mkdir(parents=True, exist_ok=True)
-            await asyncio.to_thread(
-                subprocess.run,
-                ["git", "init"],
-                cwd=target,
-                check=True,
-                capture_output=True,
-                text=True,
-            )
-            await asyncio.to_thread(
-                subprocess.run,
-                ["git", "remote", "add", "origin", args[-2]],
-                cwd=target,
-                check=True,
-                capture_output=True,
-                text=True,
-            )
-            (target / "doc.md").write_text("query credential content", encoding="utf-8")
-            return ""
-        if args == ["remote", "set-url", "origin", clean_url]:
-            assert cwd is not None
-            await asyncio.to_thread(
-                subprocess.run,
-                ["git", *args],
-                cwd=cwd,
-                check=True,
-                capture_output=True,
-                text=True,
-            )
-            return ""
-        if args == ["ls-files", "-z"]:
-            return "doc.md\x00"
-        if args == ["rev-parse", "HEAD"]:
-            return "rev-query\n"
-        return ""
-
-    monkeypatch.setattr(GitKnowledgeSource, "_run_git", _fake_run_git)
-
-    result = await refresh_knowledge_binding("docs", config=config, runtime_paths=runtime_paths)
-    key = resolve_published_index_key("docs", config=config, runtime_paths=runtime_paths)
-    metadata_text = published_index_metadata_path(key).read_text(encoding="utf-8")
-    git_config_text = (docs_path / ".git" / "config").read_text(encoding="utf-8")
-
-    assert result.index_published is True
-    assert clone_envs
-    assert "query-secret" in str(clone_envs[0])
-    assert "frag-secret" in str(clone_envs[0])
-    assert clean_url in git_config_text
-    assert "query-secret" not in git_config_text
-    assert "frag-secret" not in git_config_text
-    assert "query-secret" not in metadata_text
-    assert "frag-secret" not in metadata_text
-    assert redact_url_credentials(config.knowledge_bases["docs"].git.repo_url) == clean_url
-
-
-@pytest.mark.asyncio
 async def test_git_pull_that_changes_one_file_only_reindexes_that_file(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
@@ -8161,76 +7561,6 @@ async def test_git_pull_that_changes_one_file_only_reindexes_that_file(
         "original body 3",
         "original body 4",
         "rewritten body 2",
-    ]
-
-
-@pytest.mark.asyncio
-async def test_existing_single_branch_checkout_switches_to_new_remote_branch(tmp_path: Path) -> None:
-    """A checkout cloned for one branch should fetch and switch to another configured branch."""
-    remote_work = tmp_path / "remote-work"
-    remote_work.mkdir()
-
-    async def _git(cwd: Path, *args: str) -> None:
-        await asyncio.to_thread(
-            subprocess.run,
-            ["git", *args],
-            cwd=cwd,
-            check=True,
-            capture_output=True,
-            text=True,
-        )
-
-    await _git(remote_work, "init", "-b", "main")
-    await _git(remote_work, "config", "user.email", "tests@example.com")
-    await _git(remote_work, "config", "user.name", "MindRoom Tests")
-    (remote_work / "doc.md").write_text("main branch content", encoding="utf-8")
-    await _git(remote_work, "add", "doc.md")
-    await _git(remote_work, "commit", "-m", "main")
-    await _git(remote_work, "checkout", "-b", "release")
-    (remote_work / "doc.md").write_text("release branch content", encoding="utf-8")
-    await _git(remote_work, "commit", "-am", "release")
-    remote_bare = tmp_path / "remote.git"
-    await asyncio.to_thread(
-        subprocess.run,
-        ["git", "clone", "--bare", str(remote_work), str(remote_bare)],
-        check=True,
-        capture_output=True,
-        text=True,
-    )
-
-    docs_path = tmp_path / "checkout"
-    main_config = _config(
-        tmp_path,
-        bases={"docs": docs_path},
-        agent_bases=["docs"],
-        git_configs={"docs": KnowledgeGitConfig(repo_url=str(remote_bare), branch="main")},
-    )
-    runtime_paths = runtime_paths_for(main_config)
-    await refresh_knowledge_binding("docs", config=main_config, runtime_paths=runtime_paths)
-    main_lookup = get_published_index("docs", config=main_config, runtime_paths=runtime_paths)
-    assert main_lookup.index is not None
-    assert [document.content for document in main_lookup.index.knowledge.search("branch", max_results=5)] == [
-        "main branch content",
-    ]
-
-    release_config = _config(
-        tmp_path,
-        bases={"docs": docs_path},
-        agent_bases=["docs"],
-        git_configs={"docs": KnowledgeGitConfig(repo_url=str(remote_bare), branch="release")},
-    )
-    result = await refresh_knowledge_binding(
-        "docs",
-        config=release_config,
-        runtime_paths=runtime_paths,
-        force_reindex=True,
-    )
-    release_lookup = get_published_index("docs", config=release_config, runtime_paths=runtime_paths)
-
-    assert result.index_published is True
-    assert release_lookup.index is not None
-    assert [document.content for document in release_lookup.index.knowledge.search("branch", max_results=5)] == [
-        "release branch content",
     ]
 
 
@@ -8418,419 +7748,6 @@ async def test_refresh_scheduler_manual_reindex_runs_without_background_queue(
     await scheduler.shutdown()
 
     assert seen == [(1024, False), (5000, True)]
-
-
-@pytest.mark.asyncio
-async def test_sync_git_source_once_unchanged_head_skips_worktree_scan(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Polling an unchanged managed checkout should not scan the working tree."""
-    manager = _git_manager(tmp_path, lfs=True)
-    git_calls: list[list[str]] = []
-
-    async def _fake_ensure_git_repository(_git_config: object) -> bool:
-        return False
-
-    async def _fake_git_rev_parse(ref: str) -> str | None:
-        if ref in {"HEAD", "origin/main"}:
-            return "same"
-        return None
-
-    async def _unexpected_git_list_tracked_files() -> set[str]:
-        msg = "unchanged Git sync should not list tracked files"
-        raise AssertionError(msg)
-
-    async def _fake_run_git(args: list[str], **_: object) -> str:
-        git_calls.append(args)
-        return ""
-
-    monkeypatch.setattr(manager.git_source, "_ensure_repository", _fake_ensure_git_repository)
-    monkeypatch.setattr(manager.git_source, "_rev_parse", _fake_git_rev_parse)
-    monkeypatch.setattr(manager.git_source, "_list_tracked_files", _unexpected_git_list_tracked_files)
-    monkeypatch.setattr(manager.git_source, "_run_git", _fake_run_git)
-
-    changed_files, removed_files, updated = await manager.git_source._sync_once(manager.git_source._git_config())
-
-    assert updated is False
-    assert changed_files == set()
-    assert removed_files == set()
-    assert ["fetch", "origin", "+refs/heads/main:refs/remotes/origin/main"] in git_calls
-    assert ["lfs", "pull", "origin", "main"] in git_calls
-    assert not any(call[:3] == ["diff", "--name-only", "--no-renames"] for call in git_calls)
-
-
-@pytest.mark.asyncio
-async def test_sync_git_source_once_skips_repeated_lfs_pull_for_already_hydrated_unchanged_head(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Unchanged LFS heads should hydrate once, then reuse the persisted hydration marker."""
-    manager = _git_manager(tmp_path, lfs=True)
-    git_calls: list[list[str]] = []
-
-    async def _fake_ensure_git_repository(_git_config: object) -> bool:
-        return False
-
-    async def _fake_git_rev_parse(ref: str) -> str | None:
-        if ref in {"HEAD", "origin/main"}:
-            return "same"
-        return None
-
-    async def _fake_git_list_tracked_files() -> set[str]:
-        return {"doc.md"}
-
-    async def _fake_run_git(args: list[str], **_: object) -> str:
-        git_calls.append(args)
-        return ""
-
-    monkeypatch.setattr(manager.git_source, "_ensure_repository", _fake_ensure_git_repository)
-    monkeypatch.setattr(manager.git_source, "_rev_parse", _fake_git_rev_parse)
-    monkeypatch.setattr(manager.git_source, "_list_tracked_files", _fake_git_list_tracked_files)
-    monkeypatch.setattr(manager.git_source, "_run_git", _fake_run_git)
-
-    changed_files, removed_files, updated = await manager.git_source._sync_once(manager.git_source._git_config())
-
-    assert updated is False
-    assert changed_files == set()
-    assert removed_files == set()
-    assert ["lfs", "pull", "origin", "main"] in git_calls
-
-    hydrated_manager = _git_manager(tmp_path, lfs=True)
-    repeated_git_calls: list[list[str]] = []
-
-    async def _fake_run_git_second(args: list[str], **_: object) -> str:
-        repeated_git_calls.append(args)
-        return ""
-
-    monkeypatch.setattr(hydrated_manager.git_source, "_ensure_repository", _fake_ensure_git_repository)
-    monkeypatch.setattr(hydrated_manager.git_source, "_rev_parse", _fake_git_rev_parse)
-    monkeypatch.setattr(hydrated_manager.git_source, "_list_tracked_files", _fake_git_list_tracked_files)
-    monkeypatch.setattr(hydrated_manager.git_source, "_run_git", _fake_run_git_second)
-
-    changed_files, removed_files, updated = await hydrated_manager.git_source._sync_once(
-        hydrated_manager.git_source._git_config(),
-    )
-
-    assert updated is False
-    assert changed_files == set()
-    assert removed_files == set()
-    assert ["lfs", "pull", "origin", "main"] not in repeated_git_calls
-
-
-@pytest.mark.asyncio
-async def test_sync_git_source_once_pulls_lfs_after_reset(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """LFS-enabled repos should explicitly pull LFS objects after resetting to the remote branch."""
-    manager = _git_manager(tmp_path, lfs=True)
-    git_calls: list[list[str]] = []
-    git_envs: list[tuple[list[str], dict[str, str] | None]] = []
-
-    async def _fake_ensure_git_repository(_git_config: object) -> bool:
-        return False
-
-    async def _fake_git_rev_parse(ref: str) -> str | None:
-        if ref == "HEAD":
-            return "before"
-        if ref == "origin/main":
-            return "after"
-        return None
-
-    list_tracked_files_results = iter([{"doc.md"}, {"doc.md"}])
-
-    async def _fake_git_list_tracked_files() -> set[str]:
-        return next(list_tracked_files_results)
-
-    async def _fake_run_git(
-        args: list[str],
-        *,
-        env: dict[str, str] | None = None,
-        **_: object,
-    ) -> str:
-        git_calls.append(args)
-        git_envs.append((args, env))
-        if args[:3] == ["diff", "--name-only", "--no-renames"]:
-            return "doc.md\n"
-        return ""
-
-    monkeypatch.setattr(manager.git_source, "_ensure_repository", _fake_ensure_git_repository)
-    monkeypatch.setattr(manager.git_source, "_rev_parse", _fake_git_rev_parse)
-    monkeypatch.setattr(manager.git_source, "_list_tracked_files", _fake_git_list_tracked_files)
-    monkeypatch.setattr(manager.git_source, "_run_git", _fake_run_git)
-
-    changed_files, removed_files, updated = await manager.git_source._sync_once(manager.git_source._git_config())
-
-    assert updated is True
-    assert changed_files == {"doc.md"}
-    assert removed_files == set()
-    assert ["lfs", "pull", "origin", "main"] in git_calls
-    assert (
-        ["checkout", "--force", "-B", "main", "origin/main"],
-        {"GIT_LFS_SKIP_SMUDGE": "1"},
-    ) in git_envs
-    assert (["reset", "--hard", "origin/main"], {"GIT_LFS_SKIP_SMUDGE": "1"}) in git_envs
-
-
-@pytest.mark.asyncio
-async def test_hydrate_git_lfs_worktree_ignores_index_extension_filters(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Index extension filters must not make the Git checkout incomplete."""
-    manager = _git_manager(tmp_path, lfs=True, include_extensions=[".md", ".mdx", ".rst"])
-    git_calls: list[list[str]] = []
-
-    async def _fake_run_git(args: list[str], **_: object) -> str:
-        git_calls.append(args)
-        return ""
-
-    async def _fake_git_rev_parse(_ref: str) -> str | None:
-        return "head"
-
-    monkeypatch.setattr(manager.git_source, "_run_git", _fake_run_git)
-    monkeypatch.setattr(manager.git_source, "_rev_parse", _fake_git_rev_parse)
-
-    await manager.git_source._hydrate_lfs_worktree(manager.git_source._git_config())
-
-    assert ["lfs", "pull", "origin", "main"] in git_calls
-
-
-@pytest.mark.asyncio
-async def test_ensure_git_lfs_available_raises_clear_runtime_error(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Missing Git LFS should raise the runtime-image guidance instead of a raw git failure."""
-    manager = _git_manager(tmp_path, lfs=True)
-
-    async def _fake_run_git(args: list[str], **_: object) -> str:
-        if args == ["lfs", "version"]:
-            msg = "git: 'lfs' is not a git command"
-            raise RuntimeError(msg)
-        return ""
-
-    monkeypatch.setattr(manager.git_source, "_run_git", _fake_run_git)
-
-    with pytest.raises(RuntimeError, match="Git LFS is required for this knowledge base"):
-        await manager.git_source._ensure_lfs_available(cwd=manager.knowledge_path)
-
-
-@pytest.mark.asyncio
-async def test_ensure_git_repository_clones_lfs_repo_with_skip_smudge_env(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Initial LFS clones should hydrate even if an old hydrated-head marker matches the cloned commit."""
-    manager = _git_manager(tmp_path, lfs=True)
-    clone_envs: list[dict[str, str] | None] = []
-    git_calls: list[list[str]] = []
-    manager.git_source.lfs_hydrated_head_path.write_text("same", encoding="utf-8")
-
-    async def _fake_run_git(
-        args: list[str],
-        *,
-        cwd: Path | None = None,
-        env: dict[str, str] | None = None,
-    ) -> str:
-        _ = cwd
-        git_calls.append(args)
-        if args[0] == "clone":
-            clone_envs.append(env)
-        return ""
-
-    async def _fake_git_rev_parse(_ref: str) -> str | None:
-        return "same"
-
-    monkeypatch.setattr(manager.git_source, "_run_git", _fake_run_git)
-    monkeypatch.setattr(manager.git_source, "_rev_parse", _fake_git_rev_parse)
-
-    cloned = await manager.git_source._ensure_repository(manager.git_source._git_config())
-
-    assert cloned is True
-    assert clone_envs == [{"GIT_LFS_SKIP_SMUDGE": "1"}]
-    assert ["lfs", "pull", "origin", "main"] in git_calls
-
-
-@pytest.mark.asyncio
-async def test_run_git_redacts_credentials_in_error_message(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Git command errors should not leak embedded URL credentials."""
-    manager = _git_manager(tmp_path)
-
-    class _FailingProcess:
-        returncode = 128
-
-        async def communicate(self) -> tuple[bytes, bytes]:
-            return (
-                b"",
-                (
-                    b"fatal: unable to access "
-                    b"'https://x-access-token:secret-token@github.com/example/private.git/': "
-                    b"The requested URL returned error: 403"
-                ),
-            )
-
-    async def _fake_create_subprocess_exec(*args: object, **kwargs: object) -> _FailingProcess:
-        _ = args, kwargs
-        return _FailingProcess()
-
-    monkeypatch.setattr(asyncio, "create_subprocess_exec", _fake_create_subprocess_exec)
-
-    with pytest.raises(RuntimeError, match="Git command failed") as exc_info:
-        await manager.git_source._run_git(
-            [
-                "clone",
-                "https://x-access-token:secret-token@github.com/example/private.git",
-                "dest",
-            ],
-        )
-
-    message = str(exc_info.value)
-    assert "secret-token" not in message
-    assert "https://***@github.com/example/private.git" in message
-
-
-@pytest.mark.asyncio
-async def test_run_git_timeout_kills_subprocess_and_raises_runtime_error(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Timed out git commands should terminate the child process and raise a redacted runtime error."""
-    manager = _git_manager(tmp_path, sync_timeout_seconds=5)
-
-    class _HangingProcess:
-        returncode: int | None = None
-
-        def __init__(self) -> None:
-            self.kill_called = False
-            self.wait_called = False
-
-        async def communicate(self) -> tuple[bytes, bytes]:
-            await asyncio.Event().wait()
-            return b"", b""
-
-        def kill(self) -> None:
-            self.kill_called = True
-
-        async def wait(self) -> int:
-            self.wait_called = True
-            self.returncode = -9
-            return -9
-
-    process = _HangingProcess()
-
-    async def _fake_create_subprocess_exec(*args: object, **kwargs: object) -> _HangingProcess:
-        _ = args, kwargs
-        return process
-
-    async def _fake_wait_for(awaitable: object, **kwargs: float) -> tuple[bytes, bytes]:
-        _ = kwargs["timeout"]
-        close = getattr(awaitable, "close", None)
-        if callable(close):
-            close()
-        raise TimeoutError
-
-    monkeypatch.setattr(asyncio, "create_subprocess_exec", _fake_create_subprocess_exec)
-    monkeypatch.setattr(asyncio, "wait_for", _fake_wait_for)
-    monkeypatch.setattr(manager.git_source, "_sync_timeout_seconds", lambda: 1.0)
-
-    with pytest.raises(RuntimeError, match=r"Git command timed out after 1s: git fetch origin main"):
-        await manager.git_source._run_git(["fetch", "origin", "main"])
-
-    assert process.kill_called is True
-    assert process.wait_called is True
-
-
-@pytest.mark.asyncio
-async def test_run_git_preserves_index_lock_and_does_not_retry(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Git lock failures should surface immediately without deleting the lock file."""
-    manager = _git_manager(tmp_path)
-    repo_root = tmp_path / "repo"
-    git_dir = repo_root / ".git"
-    git_dir.mkdir(parents=True, exist_ok=True)
-    lock_path = git_dir / "index.lock"
-    lock_path.write_text("", encoding="utf-8")
-
-    class _FailingProcess:
-        returncode = 128
-
-        async def communicate(self) -> tuple[bytes, bytes]:
-            return (
-                b"",
-                (
-                    f"fatal: Unable to create '{lock_path}': File exists.\n"
-                    "Another git process seems to be running in this repository."
-                ).encode(),
-            )
-
-    recorded_cwds: list[str] = []
-
-    async def _fake_create_subprocess_exec(*args: object, **kwargs: object) -> object:
-        _ = args
-        recorded_cwds.append(str(kwargs["cwd"]))
-        return _FailingProcess()
-
-    monkeypatch.setattr(asyncio, "create_subprocess_exec", _fake_create_subprocess_exec)
-
-    with pytest.raises(RuntimeError, match=r"index\.lock"):
-        await manager.git_source._run_git(["checkout", "main"], cwd=repo_root)
-
-    assert recorded_cwds == [str(repo_root)]
-    assert lock_path.exists() is True
-
-
-@pytest.mark.asyncio
-async def test_run_git_cancellation_kills_subprocess(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Cancelling a git command should terminate and reap the child process."""
-    manager = _git_manager(tmp_path)
-    wait_forever = asyncio.Event()
-
-    class _HangingProcess:
-        returncode: int | None = None
-
-        def __init__(self) -> None:
-            self.kill_called = False
-            self.wait_called = False
-
-        async def communicate(self) -> tuple[bytes, bytes]:
-            await wait_forever.wait()
-            return b"", b""
-
-        def kill(self) -> None:
-            self.kill_called = True
-
-        async def wait(self) -> int:
-            self.wait_called = True
-            self.returncode = -9
-            return -9
-
-    process = _HangingProcess()
-
-    async def _fake_create_subprocess_exec(*args: object, **kwargs: object) -> _HangingProcess:
-        _ = args, kwargs
-        return process
-
-    monkeypatch.setattr(asyncio, "create_subprocess_exec", _fake_create_subprocess_exec)
-
-    task = asyncio.create_task(manager.git_source._run_git(["fetch", "origin", "main"]))
-    await asyncio.sleep(0)
-    task.cancel()
-
-    with pytest.raises(asyncio.CancelledError):
-        await task
-
-    assert process.kill_called is True
-    assert process.wait_called is True
 
 
 def test_redact_url_credentials_hides_entire_http_userinfo() -> None:
